rename fuzz_cov_load to cov_load
